Migrate V2FramePage from JS to TypeScript with improved types and styling
